Add redis deserialization back (not using the deep deserialization)
Structs in model no only hold one level of information. The rest can
be retrieved from Redis as needed. This allows GraphQL to be more lazy
in loading data, and removes the circular deserialization loop which
was a problem before. Some more endpoints need to be implemented, but
this starts the work.

Also, looking at the implementation of FromID, we could probably make
a custom derive of it if the struct already implements Deserialize and
has the key() method (would need a trait for that as well). The issue is
that custom derives require nightly Rust atm and would need to be in
its own crate. Or, the logic could be generified and made into a
separate method.

Signed-off-by: Mcat12 <newtoncat12@yahoo.com>