Only Store Binaries In Workspace
Updating the workspace was taking around 3m. This should improve build
time significantly by only updating the cache when dependencies change,
and only using the final binaries in the integration tests.

This breaks running the integration tests locally with the `circleci`
tool. They can still be run locally, just not with the tool.

Closes #27