Added Actix and Implemented GraphQL Endpoints
I followed the example in the actix examples repository.
https://github.com/actix/examples/blob/4d41d40da2d73579867ab02cf898a176f9494795/juniper/src/main.rs

I considered using sync handlers since everything is sync anyway but
decided against it because from what I understand of tokio and futures,
there are less threads than connections. The threads will be burdened
and may be slower if more work is being done on them, slowing down other
connection handling. This might be a pre-mature optimization (node
handles 1000s of connections on a single thread), but I'm not
comfortable enough with this library and paragidm to experiment yet.
Also, when I tried doing it, I had some issues with the error type.