Purge CI Cache Whenever It Changes
Changed ci to not inherit past cached assets. The cache is growing large
(~2.5 GB) and slowing the build down. Most of that is probably old
versions of libraries which aren't used anymore. The cache should be
purged on cache updates. This will slow down the buidls for cahce
misses, but improve the performance of builds for cache hits.

This has reduced the cache size to around 500MB build time to ~3m.
https://circleci.com/gh/forte-music/core/350