Implemented RangeStream
It allows for streaming a from an io.Read + io.Seek using range
requests. See the documentation and code for more.

I plan to commit this back to actix_web soon. I don't know how it will
be received because there is a very similar implementation `NamedFile`
which has already been released. Making big changes to it would be a
breaking change.