Added Initial Transcoding Support
Primitive eager transcoding of a audio file has been implemented.  This
was both difficult and rewarding to implement. There are a couple of
memory management kinks which should be worked out during the next step.

The ffmpeg create is capable of being statically linked. This dependency
really slows down compilation time.

See #9