Fixed Path Handling
Breaks Windows support. Windows support isn't a priority. Run it inside
a VM or Ubuntu on Windows on Linux if you really want.

Closes #63

File paths are not necessarily valid strings, so they should be stored
as an array of bytes instead of a valid string in the database.

On Unix OsStr and OsString map to an array of bytes. On Windows they map
to an array of u16s. There isn't an easy way to store an array of u16s
in SQLite. Also storing u16s on Windows and u8s on Linux would break
database portability. I don't think we care but implementing
serialization of u16s is annoying.